requirements.txt
customtkinter
pypresence
requests
beautifulsoup4
lxml
Pillow
